+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
  - Fixed IR Pulse Component to pulse for the correct pulse time, instead of the total duration
+ - Removed SafeCriticalSection.TryEnter from HttpPort.Standard, FeedbackDebounce, Heartbeat, and AbstractSerialBuffer
 
 ## [14.1.0] - 2021-08-18
 ### Added